Deleted alert.jsx and button.jsx from atom dir
