+++ components/atom/button2.jsx
@@ -1,6 +1,45 @@
 import { Button } from "react-bootstrap";
 import { img } from "./image";
 
+export const sign_in = (signIn) => {
+  return (
+    <Button
+      className="justify-content-right"
+      variant="outline-primary"
+      onClick={() => signIn()}
+    >
+      Sign In
+    </Button>
+  );
+};
+
+export const sign_out = (signOut) => {
+  return (
+    <Button
+      className="justify-content-right"
+      variant="outline-success"
+      onClick={() =>
+        signOut({
+          callbackUrl: `${window.location.origin}/dashboard`,
+        })
+      }
+    >
+      Sign Out
+    </Button>
+  );
+};
+
+export const RadioButton = ({ checked, onChange }) => (
+  <>
+    <input
+      type="radio"
+      onChange={onChange}
+      name="radio"
+      defaultChecked={checked}
+    />
+  </>
+);
+
 export const submitButton = (handleSubmit, Name, variant) => {
   return (
     <div className="form-group" style={{ padding: "10px" }}>
